tests/monitor-abstraction: Assign non-conflicting CRTCs for tiled monitors
The preferred-mode and changed-connector tests would have conflicting CRTCs
between the normal monitor and output for the primary tile of the tiled
monitor in the tests. This change removes the conflict to prevent failures
in the meta_monitor_config_manager_assign path.

Signed-off-by: Rahul Rameshbabu <rrameshbabu@nvidia.com>
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4685>